+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for path based color automation.
 - Added `window.decorations` property, available options are `Enabled`, `Disabled`, `Transparent` and `Buttonless`.
+- Removed `macos-hide-toolbar-buttons` in favor of `window.decorations` api.
 
 ## 0.0.32